chore(refactor): refactored debounce, Recipes, and fetch
