+++ src/utils/main.ts
@@ -1,2 +1,3 @@
 export { default as fetchID } from './fetchID';
 export { default as fetchFormat } from './fetchFormat';
+export { default as ingredientSearch } from './queryStringBuilder';
